refactor(structure): migrate to domain-based folders and configure @ alias (WIP)
- Restructured folders (context, components, hooks, etc.) into domain-based structure to simplify imports.
- Configured `@` alias in Vite config.
- Next steps:
  - Fix `FollowedBy` to display "followed by you" if logged user follows the profile.
  - Fix counters navigation.
