+++ src/components/auth/index.js
@@ -1,0 +1,1 @@
+export { default as AuthForm } from "./AuthForm";
