+++ src/components/chat/index.js
@@ -1,0 +1,1 @@
+export { default as Messages } from "./Messages";
