+++ src/components/navigation/index.js
@@ -1,0 +1,2 @@
+export { default as Navbar } from "./Navbar";
+export { default as PrivateRoute } from "./PrivateRoute";
