+++ src/components/post/buttons/index.js
@@ -1,0 +1,1 @@
+export { default as PostActionRow } from "./PostActionRow";
