+++ src/components/post/index.js
@@ -1,0 +1,7 @@
+export { default as AddComment } from "./AddComment";
+export { default as CommentsItem } from "./CommentsItem";
+export { default as NewPostForm } from "./NewPostForm";
+export { default as PostImage } from "./PostImage";
+export { default as PostItem } from "./PostItem";
+export { default as PostList } from "./PostList";
+export * from "./buttons"; // 👈 exporta los de /buttons
