+++ src/components/user/index.js
@@ -1,0 +1,5 @@
+export { default as ProfileStats } from "./ProfileStats";
+export { default as UserFollowedBy } from "./UserFollowedBy";
+export { default as UserList } from "./UserList";
+export { default as UserProfileSidebar } from "./UserProfileSidebar";
+export { default as UserSuggestions } from "./UserSuggestions";
