+++ src/hooks/auth/useAuth.js
@@ -1,6 +1,6 @@
 //Hook para acceder al contexto.
 import { useContext } from "react";
-import { AuthContext } from "../context/AuthContext";
+import { AuthContext } from "@/context";
 
 export function useAuth() {
     const context = useContext(AuthContext);
