+++ src/hooks/users/useFollowers.js
@@ -1,5 +1,5 @@
 // hooks/useFollowers.js
-import { useUserConnections } from "./useUserConnections";
+import { useUserConnections } from "@/hooks/users";
 
 export function useFollowers(userId, initialPage = 1) {
   return useUserConnections("followers", userId, initialPage);
