+++ src/hooks/users/useFollowing.js
@@ -1,5 +1,5 @@
 // hooks/useFollowing.js
-import { useUserConnections } from "./useUserConnections";
+import { useUserConnections } from "@/hooks/users";
 
 export function useFollowing(userId, initialPage = 1) {
   return useUserConnections("following", userId, initialPage);
