+++ src/main.jsx
@@ -3,8 +3,8 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ErrorBoundary } from "react-error-boundary";
+import { AuthProvider } from "@/context";
 import App from './App.jsx'
-import { AuthProvider } from './context/AuthContext.jsx';
 import "./styles/index.css";
 
 function Fallback({ error }) {
