+++ src/pages/Home.jsx
@@ -1,8 +1,7 @@
 // src/components/Home.jsx
 import { useContext } from "react";
-import { AuthContext } from "../context/AuthContext";
-import PostList from "../components/post/PostList";
-import NewPostForm from "../components/post/NewPostForm";
+import { AuthContext } from "@/context";
+import { PostList, NewPostForm  } from "@/components/post";
 
 function Home() {
     const { user } = useContext(AuthContext);
