+++ src/pages/index.js
@@ -1,0 +1,7 @@
+export { default as AuthPage } from "./AuthPage";
+export { default as EditUserProfile } from "./EditUserProfile";
+export { default as Home } from "./Home";
+export { default as MainLayout } from "./MainLayout";
+export { default as People } from "./People";
+export { default as Timeline } from "./Timeline";
+export { default as UserPage} from "./UserPage";
